refactor(api): group auth under lib/auth
Splits the better-auth instance from the Elysia macro, so any controller can
require a session with `{ auth: true }`, and keeps the handler on a route
under the path it serves rather than a prefix mount — mounting strips the
prefix, which would leave better-auth generating callback URLs missing `/api`.
