+++ apps/api/src/lib/auth/better-auth.ts
@@ -4,17 +4,16 @@ import { sql } from '#db/client.ts';
 import { env } from '#lib/env.ts';
 import { RoutePrefix } from '#lib/routes/prefixes.ts';
 
-// Created by src/db/migrations/0002_create_auth_schema.sql — the two have to
-// name the same schema.
 const AUTH_SCHEMA = 'auth';
 
 /** Bare path; the api controller applies the `/api` prefix it sits under. */
 export const AUTH_ROUTE_PATH = '/auth';
+const AUTH_BASE_PATH = `${RoutePrefix.Api}${AUTH_ROUTE_PATH}`;
 
 export const auth = betterAuth({
   database: bunSqlAdapter({ sql, pgSchema: AUTH_SCHEMA }),
   baseURL: env.BASE_URL.origin,
-  basePath: `${RoutePrefix.Api}${AUTH_ROUTE_PATH}`,
+  basePath: AUTH_BASE_PATH,
   secret: env.BETTER_AUTH_SECRET,
   socialProviders: {
     github: {
