+++ apps/api/src/routes/api/auth/controller.ts
@@ -1,5 +1,5 @@
 import { Elysia } from 'elysia';
-import { AUTH_ROUTE_PATH, auth } from '#lib/auth.ts';
+import { AUTH_ROUTE_PATH, auth } from '#lib/auth/better-auth.ts';
 
 // better-auth owns every path under its base path and reads the request itself,
 // so Elysia must not consume the body before handing it over.
